Linux hack improved.
By setting LD_DEBUG=libs before executing Urho3DPlayer we can see that
the very first time we try to unload the library, the system doesn't
really unload it (there's an "init" but not a "fini" associated).
However, the rest of the times the libraries are unloaded as they
should. So, we only change the library name the first time.